Fix extremely bad channeling bug (test notes in desc.)
Test suites were not picking up on this bug because they use on-the-fly skills rather than cooldowns. I'm not sure whether  that means the priority for addressing this issue going forward should be a) finding a way to add in cooldowns and charges to the test suites, b) creating unit tests for individual functions like Requirement.usable somehow, or c) just run a manual integration test from time to time, which caught this bug immediately. This is the only time in the history of the project that something like this has happened.